refactor: 홈 레이아웃 순서 변경
