+++ components/ProgressBar.tsx
@@ -1,6 +1,5 @@
 'use client';
-import { useEffect, useRef } from 'react';
-import { motion, useInView, useAnimationControls } from 'framer-motion';
+import { motion, useAnimationControls } from 'framer-motion';
 type ProgressBarProps = {
   progressPercent: number;
   barColor: 'primary' | 'white';
@@ -8,18 +7,6 @@ type ProgressBarProps = {
 
 export const ProgressBar = ({ progressPercent, barColor }: ProgressBarProps) => {
   const controls = useAnimationControls();
-  const ref = useRef(null);
-  const isInview = useInView(ref);
-
-  useEffect(() => {
-    controls.set({ width: '0%' });
-
-    if (isInview) {
-      controls.start({ width: `${progressPercent}%` });
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isInview]);
 
   const barColorClass = {
     primary: 'bg-primary',
@@ -28,10 +15,15 @@ export const ProgressBar = ({ progressPercent, barColor }: ProgressBarProps) => 
 
   return (
     <motion.div
-      ref={ref}
       className={`h-full w-full ${barColorClass[barColor]} rounded-full`}
       initial={{
         width: '0%'
+      }}
+      whileInView={{
+        width: `${progressPercent}%`
+      }}
+      viewport={{
+        once: true
       }}
       animate={controls}
       transition={{ duration: 0.25 }}
